device/*: Use clock_hwclock mixin
This reduces needless duplication, and prepares the terrain for better
re-use down the line, and more importantly, feature detection.

One thing that should be done is have a generic "clock_automatic" mixin
that peeks at the environment and determines whether to use timedatectl,
hwclock, or other *standard* methods.

This does not prevent platforms from implementing platform-specific
methods of handling the time. See platforms that still use `hwclock`,
but in a different manner.