device/sdl: Extend any device with distribution method specifics
This makes the logic for appimage or ubuntu touch versions apply on top
of any detected device.

This, in turn, means that device-specific handling will be respected as
defined for the given device.

For example, running an AppImage on the PineNote *should* allow the
proper device to be detected.

Whether or not all features will work is another issue. I don't have the
required knowledge to know if, for example, `light` will work for
setting the backlight with the other backends.

The logic, though is good enough that if it is a problem, we can change
the backends as needed to work under the different distribution methods.