[WIP] Add “select a network” menu for Wi-Fi
This allows listing the connections directly, without requiring the user
to turn off wireless networking beforehand.

This is mainly a UX improvement, as it can be confusing to have to futz
about with the toggle when you want to connect to a network. Wireless
networking is already enabled, it's not obvious how to *just* connect to
a network.