frontend/ui: Allow removing Exit from menu
There are scenarios where KOReader is the only expected frontend
application in-used on the device.

Namely, using KOReader running as the sole system on a mainline
Linux-based device.

In this situation, it makes no sense to "Exit" KOReader.

This also reworks the logic for reducing the menu to a single item. It
will only happen when `Exit` is the only active item.