device/mixins: Add `light`-backed powerd mixin for frontlight
This implements a naïve algorithm to mix the cool and warm backlights.
This is sufficient for the only system this has been tested with, but
possibly needs some thought.

Why is this converting `/sys/` paths instead of being configured
directly with the light-compatible identifiers? It's so the *knowledge*
about the system (the backlight node names) stays static, whichever
backend is used. In turn, this should allow drop-in replacement of the
backends.