device/sdl: Add preliminary smarter power button handling
For now it only suspends; long press being ignored.

It is expected that with time that a long press would be used to show
a quick menu with power options.

Note that this is purposefully built in a way where user patches can be
used. I have a personal use-case with a WIP device where sometimes
calling a screen refresh would be useful, but impossible given technical
limitations.

With the way that this is built, I could replace the short power button
press with a screen refresh ioctl, and add suspend back to long press in
a user patch. (Not verified to work at this moment.)