datastorage: Use util.makePath for parent directories
When launching KOReader for the first time on a system where *nothing
else* used `$XDG_CONFIG_HOME`, KOReader would crash once.

With this change, it won't crash anymore, since the directory structure
is present.

The assumption that parent directories exist may not hold true when
running an "appliance" style deployment where KOReader is the only thing
in use on the system.