XXX: Work around resize bug at first launch with SDL ...
... but this is likely not the proper fix.

Technically this should fix the same rendering bug on any SDL builds
when starting from the last document too.

TLDR; filemanager never inhibits input at launch, so it doesn't exhibit
the issue. readerui inhibits at launch, resize events will have been
queued early.

If the window size is different than the saved (or default) dimensions,
the rendering will be for the saved value, in a window of a different
size.

This is especially bad in a system where the window size is not
modifiable, like an "appliance" style deployment of KOReader.

Another alternative to this fix is to never inhibit handleSdlEv.

Finally, the better alternative would be to add a new even category, and
use it for "system level" events. I don't know if any other events than
those for window management would fall in that category.

Under some appliance-style environments, an alternative would be to use
`SDL_FULLSCREEN`. It may work around the issue more gracefully by using
the intrinsic display sizes.